refactor: sever와 app 분리
