+++ index.js
@@ -1,7 +1,7 @@
 require("dotenv").config();
 
 const express = require("express");
-const initAsyncApp = require("./src/loaders/index");
+const initAsyncApp = require("./src/loaders/initAsyncApp");
 const loadHttpServer = require("./src/loaders/server");
 
 loadHttpServer(initAsyncApp(express()));
